refactoring GroupHeaderOptions component
