+++ src/components/group/GroupHeader.jsx
@@ -13,9 +13,9 @@ const GroupHeaderInfo = ({ groupName = 'viaje brasil', cantPersons = 1 }) => {
 
     return (
         <>
-            <div className='flex items-center w-96 justify-center gap-5 mt-6'>
+            <div className='flex items-center justify-center gap-5 mt-6'>
                 <Avatar color='primary'>{groupName?.charAt(0).toUpperCase()}</Avatar>
-                <div className='flex flex-col items-start gap-2'>
+                <div className='flex flex-col items-start gap-2 '>
                     <h2 className='m-0 capitalize'>{groupName}</h2>
                     <GroupHeaderOtions />
                 </div>
